pyproject.toml
[tool.poetry]
name = "binarybattle"
version = "0.1.0"
description = ""
authors = ["Samama251251 <samama251251@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
channels = "^4.1.0"
channels-redis = "^4.2.0"
daphne = "^4.1.2"
uvicorn = {extras = ["standard"], version = "^0.32.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
